Column re-alignment of 2002 and 2006 datasets.
I intended to parse this with a regex and avoid touching the original
data files, but it was either imperfect with a couple of false +ves or
too complicated accounting for these exceptions. I have therefore
resorted to ensuring there is 2+ <space> between each column (field
strings have <=1 <space> internally).

This is not the end of the world. The files are diff-able and a word
diff shows that none of the strings have been altered. It makes the
parsing regex a lot less convoluted and therefore simplifies the code.