Introduce a constants property to the parser.
Contains a list of physical constants derived from the table records.
The derivation involves some string replacement (to remove
cruft/non-parsable characters etc) and some type coercion. These are
stored in a named tuple the same as the records but of a
"PhysicalConstant" type and not "CODATARecord" type purely for
semantic reasons.